refactor: extract cards and effects
