+++ backend/game/cards.js
@@ -1,0 +1,16 @@
+const CARD_LIST = [
+  { id: 1, name: "兵士", enName: "soldier", cost: 1, count: 5 },
+  { id: 2, name: "道化", enName: "clown", cost: 2, count: 2 },
+  { id: 3, name: "騎士", enName: "knight", cost: 3, count: 2 },
+  { id: 4, name: "僧侶", enName: "monk", cost: 4, count: 2 },
+  { id: 5, name: "魔術師", enName: "sorcerer", cost: 5, count: 2 },
+  { id: 6, name: "将軍", enName: "general", cost: 6, count: 1 },
+  { id: 7, name: "大臣", enName: "minister", cost: 7, count: 1 },
+  { id: 8, name: "姫", enName: "princess", cost: 8, count: 1 },
+  { id: 9, name: "姫(眼鏡)", enName: "princess_glasses", cost: 8, count: 1 },
+  { id: 10, name: "伯爵夫人", enName: "countess", cost: 8, count: 1 },
+  { id: 11, name: "女侯爵", enName: "marchioness", cost: 7, count: 1 },
+  { id: 12, name: "姫(爆弾)", enName: "princess_bomb", cost: 8, count: 1 },
+];
+
+module.exports = CARD_LIST;
